Enable the sidebar and move "back" button to the key binding overlay
